LPS-75321 Moves code to LayoutsAdminDisplayContext
